requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
termcolor==2.4.0